+++ apps/router-demo/router-host-2000/src/pages/Detail.tsx
@@ -1,0 +1,19 @@
+import { loadRemote } from '@module-federation/enhanced/runtime';
+// import Button from 'remote1/button';
+
+// console.log('loadRemote', loadRemote);
+// loadRemote("remote1/button").then((m) => {
+// 	console.log(m);
+// });
+
+function Detail() {
+  return (
+    <div>
+      Detail
+      <h5>Remote Button</h5>
+      {/* <Button /> */}
+    </div>
+  );
+}
+
+export default Detail;