+++ apps/router-demo/router-remote2-2002/src/export-App.tsx
@@ -1,0 +1,12 @@
+import App from './App';
+import { createBridgeComponent } from '@module-federation/bridge-react';
+
+export const provider = createBridgeComponent({
+  rootComponent: App,
+});
+// // console.log('xxxx', createBridgeComponent(App));
+
+// export default {
+//     provider: createBridgeComponent(App),
+// }
+// export default provider;