+++ packages/rspack/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @module-federation/rspack
 
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies [804cc3b]
+- Updated dependencies [52bb94c]
+- Updated dependencies [d2ab821]
+- Updated dependencies [f7a6d3f]
+  - @module-federation/dts-plugin@0.2.0
+  - @module-federation/bridge-react-webpack-plugin@0.2.0
+  - @module-federation/manifest@0.2.0
+  - @module-federation/sdk@0.2.0
+  - @module-federation/runtime-tools@0.2.0
+  - @module-federation/managers@0.2.0
+
 ## 0.1.21
 
 ### Patch Changes